all refactoring ver2

+++ src/board/posts/services/post-command.service.ts
@@ -1,21 +1,18 @@
-import { BadRequestException, Injectable, InternalServerErrorException, NotFoundException, UnauthorizedException } from '@nestjs/common';
-
-//data
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+//model
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { SortOrder } from 'mongoose';
 import { Post, PostDocument } from '../schemas/post.schema';
 import { PostLike, PostLikeDocument } from '../schemas/post-like.schema';
-
 //dto
-import { PostCommandDTO } from '../dto/req/post-command.dto';
+import { CreatePostDto, UpdatePostDto } from '../dto/req/post-command.dto';
 
 @Injectable()
 export class PostCommandService {
   constructor(
     @InjectModel(Post.name) private readonly postModel: Model<PostDocument>,
     @InjectModel(PostLike.name) private readonly postLikeModel: Model<PostLikeDocument>,
-  ) { }
+  ) {}
 
   /**게시물 좋아요 동작*/
   async postLike(postId: string, userId: string): Promise<'liked' | 'unliked'> {
@@ -29,7 +26,7 @@ export class PostCommandService {
   }
 
   /**게시물 작성*/
-  async createPost(dto: PostCommandDTO.Create, userId: string): Promise<string> {
+  async createPost(dto: CreatePostDto, userId: string): Promise<string> {
     const created = new this.postModel({
       ...dto,
       authorId: userId,
@@ -40,8 +37,12 @@ export class PostCommandService {
   }
 
   /**게시물 업데이트*/
-  async updatePost(dto: PostCommandDTO.Update, postId: string): Promise<boolean> {
-    const result = await this.postModel.updateOne({ _id: postId }, { $set: dto }, { runValidators: true, timestamps: true });
+  async updatePost(dto: UpdatePostDto, postId: string): Promise<boolean> {
+    const result = await this.postModel.updateOne(
+      { _id: postId },
+      { $set: dto },
+      { runValidators: true, timestamps: true },
+    );
     if (result.matchedCount === 0) throw new NotFoundException('해당 게시물 없음');
     if (result.modifiedCount === 0) throw new BadRequestException('변경 된 내용 없음');
 
@@ -50,10 +51,15 @@ export class PostCommandService {
 
   /**게시물 및 게시물 좋아요 삭제*/
   async selfDeletePost(postId: string, deletedAt: Date, deletedBy: string): Promise<boolean> {
-    await Promise.all([
-      this.postModel.updateOne({ _id: postId }, { $set: { isDeleted: true, likecount: 0, deletedAt: deletedAt, deletedBy: deletedBy } }),
-      this.postLikeModel.deleteMany({ postId })
-    ]);
+    const updated = await this.postModel.updateOne(
+      { _id: postId },
+      { $set: { isDeleted: true, likecount: 0, deletedAt: deletedAt, deletedBy: deletedBy } },
+    );
+    if (updated.matchedCount === 0) throw new NotFoundException('해당 게시물 없음');
+    if (updated.modifiedCount === 0) throw new BadRequestException('변경 된 내용 없음');
+
+    await this.postLikeModel.deleteMany({ postId });
+
     return true;
   }
 
@@ -61,7 +67,7 @@ export class PostCommandService {
   async createLike(postId: string, userId: string): Promise<'liked'> {
     await Promise.all([
       this.postLikeModel.create({ postId, userId }),
-      this.postModel.findByIdAndUpdate(postId, { $inc: { likeCount: 1 } }),
+      this.postModel.updateOne({ _id: postId }, { $inc: { likeCount: 1 } }),
     ]);
     return 'liked';
   }
@@ -70,7 +76,7 @@ export class PostCommandService {
   async deleteLike(postId: string, userId: string): Promise<'unliked'> {
     await Promise.all([
       this.postLikeModel.deleteOne({ postId, userId }),
-      this.postModel.findByIdAndUpdate(postId, { $inc: { likeCount: -1 } }),
+      this.postModel.updateOne({ _id: postId }, { $inc: { likeCount: -1 } }),
     ]);
     return 'unliked';
   }
